tt evaluate
