+++ Chess98/base.hpp
@@ -14,7 +14,6 @@
 
 struct Piece;
 struct Move;
-struct TtItem;
 using TRICK_RET = std::pair<bool, int>;
 using SEARCH_RET = std::pair<Move, int>;
 using POS = int;
@@ -100,21 +99,6 @@ struct Move {
     const POS beg { 0 };
     const POS end { 0 };
     MOVE_TYPE type { NORMAL };
-};
-
-struct TtItem {
-    TtItem() = default;
-
-    int hash_lock { 0 };
-    int vlExact { 0 };
-    int vlBeta { 0 };
-    int vlAlpha { 0 };
-    int exactDepth { 0 };
-    int betaDepth { 0 };
-    int alphaDepth { 0 };
-    Move exact_move {};
-    Move beta_move {};
-    Move alpha_move {};
 };
 
 namespace utils {
